+++ app/adapters/application.js
@@ -1,7 +1,5 @@
 import JSONAPIAdapter from '@ember-data/adapter/json-api';
 import { service } from '@ember/service';
-// eslint-disable-next-line ember/no-computed-properties-in-native-classes
-import { computed } from '@ember/object';
 import ENV from 'houseninja/config/environment';
 
 export default class ApplicationAdapter extends JSONAPIAdapter {
@@ -9,14 +7,11 @@ export default class ApplicationAdapter extends JSONAPIAdapter {
 
   host = ENV.apiHost;
 
-  // eslint-disable-next-line
-  @computed('session.{data.authenticated.access_token,isAuthenticated}')
   get headers() {
-    let headers = {};
     if (this.session.isAuthenticated) {
-      // eslint-disable-next-line
-      headers['Authorization'] = `Bearer ${this.session.data.authenticated.access_token}`;
+      return this.session.authenticatedHeaders;
+    } else {
+      return {};
     }
-    return headers;
   }
 }